Created and Tested a mean-variance-standard-deviation-calculator
The input of the function has a list containing 9 digits. The function  converts the list into a 3 x 3 Numpy array, and then returns a dictionary containing the mean, variance, standard deviation, max, min, and sum along both axes and for the flattened matrix.
If a list containing less than 9 elements is passed into the function, it raises a ValueError exception with the message: "List must contain nine numbers." The values in the returned dictionary are not lists and not Numpy arrays.

The unit tests for this project are in test_module.py.

This is a part of free-code-camp data-analysis projects